+++ servers/game/src/main/kotlin/fr/rob/game/domain/world/UpdatableInterface.kt
@@ -1,5 +1,5 @@
 package fr.rob.game.domain.world
 
 interface UpdatableInterface {
-    fun update(deltaTime: Long)
+    fun update(deltaTime: Int)
 }
